Start migrating stuff to a Query class [SERINA-10]
This allows a relation object graph to be built, to see which objects
(once hydrated) belong to which other object. Previously attempted this
directly with the mappers, but since mappers are a distributed system,
there was no convenient way to manufacture this object graph. A way to
arbitrarily get the right references etc was needed without getting
messy. Thus, class Query

This should be able to arbitrarily reach out to various mapper classes
based on the select() definition by concatting the model name with
Mapper.