Store final joined collections in specified field [SERINA-10]
In rules, you can now specify a [$this][collection] field, which is
where stuff from [$other] is placed. This is used to counter the
scenario where table1 joins table2 on a key stored from table2, ie.

join on table1.id = table2.user_id

That is, table1 never knew about table2.

Prior to this, $table1 had a foreign key for items in $table2 (and the
latter didn't know about the former)