Tried many-many relations [SERINA-10]
I thought this would allow me to leverage all the existing work, sliding
in a "using" rule element. But then I realised that there was no way to
know which objects related to which other objects in this context
without having to go through a lot of hoops. The relationship between
the two primary keys (e.g. user_id and phone_id) are stored in a
relation table which has no mappers or models, nor is it part of the
overall select() statement.

For now it might actually be easier (and more convenient?) to do it
manually, or with multiple database calls instead.