Invert joining and relations [SERINA-11]
Instead of joining on columns, Mappers now join on properties. This
relieves the need to have columns defined (such as address) when they
will be used as hydrated Collections from other tables. In this case, a
column would always remain null, but still be in use in the Mapper and
join definitions, forcing it to exist.

Inverting the joins and relations now allows Mappers to join on
Properties, rather than Columns.